refactor code

+++ src/tutors/referees/referees.service.ts
@@ -1,4 +1,8 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  ForbiddenException,
+  Injectable,
+} from '@nestjs/common';
 import { Prisma } from '@prisma/client';
 import { PrismaService } from 'src/prisma.service';
 
@@ -35,19 +39,44 @@ export class RefereesService {
     );
   }
 
-  async update(id: number, updateRefereeDto: Prisma.RefereesUpdateInput) {
-    return await this.prisma.referees.update({
-      where: { id },
-      data: updateRefereeDto,
-    });
-  }
-
-  async remove(id: number) {
+  async update(
+    id: number,
+    tutorId: number,
+    updateRefereeDto: Prisma.RefereesUpdateInput,
+  ) {
     try {
-      await this.prisma.referees.delete({ where: { id } });
-      return { message: 'Referee deleted' };
+      if (await this.checkReferee(id, tutorId)) {
+        return await this.prisma.referees.update({
+          where: { id },
+          data: updateRefereeDto,
+        });
+      }
     } catch (error) {
       throw new BadRequestException('Referee not found');
     }
+  }
+
+  async remove(id: number, tutorId: number) {
+    try {
+      if (await this.checkReferee(id, tutorId)) {
+        await this.prisma.referees.delete({ where: { id } });
+        return { message: 'Referee deleted' };
+      }
+    } catch (error) {
+      throw new BadRequestException('Referee not found');
+    }
+  }
+
+  private async checkReferee(
+    refereeId: number,
+    tutorId: number,
+  ): Promise<Boolean> {
+    const referee = await this.prisma.referees.findUnique({
+      where: { id: refereeId },
+    });
+    if (+referee.tutorId !== +tutorId) {
+      throw new ForbiddenException('You are not the owner of this referee');
+    }
+    return true;
   }
 }
